Add serialization for ReplayFrame Change ReplayMutator file handling Move some replay stuff around
